refactor(suite): remove final onboarding step
