+++ packages/suite-desktop-core/e2e/tests/onboarding/analytics-consent.test.ts
@@ -26,7 +26,7 @@ test.describe(
                 await onboardingPage.enterTHPPairingCode();
             }
 
-            await onboardingPage.onboardingContinueButton.click();
+            await onboardingPage.onboardingExitButton.click();
 
             if (model.isModelWithSecureElement()) {
                 await onboardingPage.passThroughAuthenticityCheck();
