+++ packages/suite-desktop-core/e2e/tests/onboarding/t1b1/t1b1-recovery-success.test.ts
@@ -55,9 +55,7 @@ test.describe(
                 // Finalize recovery, skip pin, and verify success
                 await onboardingPage.continueRecoveryButton.click();
                 await onboardingPage.pin.skip();
-                await onboardingPage.continueCoinsButton.click();
-                await expect(onboardingPage.finalTitle).toBeVisible();
-                await expect(onboardingPage.finalTitle).toContainTranslation('TR_FINAL_HEADING');
+                await expect(onboardingPage.onboardingExitButton).toBeVisible();
             },
         );
     },
