+++ packages/suite/src/views/onboarding/steps/BasicSettings/index.tsx
@@ -1,8 +1,9 @@
-import { selectEnabledNetworks } from '@suite-common/wallet-core';
+import { selectEnabledNetworks, startDiscoveryThunk } from '@suite-common/wallet-core';
+import { EventType, analytics } from '@trezor/suite-analytics';
 
 import { OnboardingCard } from 'src/components/onboarding/OnboardingCard/OnboardingCard';
 import { Translation } from 'src/components/suite/Translation';
-import { useOnboarding, useSelector } from 'src/hooks/suite';
+import { useDevice, useDispatch, useOnboarding, useSelector } from 'src/hooks/suite';
 import { getIsTorLoading } from 'src/utils/suite/tor';
 
 import { BasicSettingsStepBox } from './BasicSettingsStepBox';
@@ -10,10 +11,38 @@ import { BasicSettingsStepBox } from './BasicSettingsStepBox';
 const BasicSettings = () => {
     const enabledNetworks = useSelector(selectEnabledNetworks);
     const torStatus = useSelector(state => state.suite.torStatus);
-    const { goToNextStep } = useOnboarding();
+    const onboardingAnalytics = useSelector(state => state.onboarding.onboardingAnalytics);
+    const { goToSuite } = useOnboarding();
+    const { device } = useDevice();
+    const dispatch = useDispatch();
 
     const noNetworkEnabled = !enabledNetworks.length;
     const isTorLoading = getIsTorLoading(torStatus);
+
+    if (device?.features === undefined) {
+        return null;
+    }
+
+    const reportAnalytics = () => {
+        const payload = {
+            ...onboardingAnalytics,
+            duration: Date.now() - onboardingAnalytics.startTime!,
+            device: device.features.internal_model,
+            unitPackaging: device.features.unit_packaging ?? 0,
+        };
+        delete payload.startTime;
+
+        analytics.report({
+            type: EventType.DeviceSetupCompleted,
+            payload,
+        });
+    };
+
+    const handleGoToSuite = () => {
+        reportAnalytics();
+        goToSuite();
+        dispatch(startDiscoveryThunk({ device }));
+    };
 
     return (
         <BasicSettingsStepBox
@@ -21,9 +50,9 @@ const BasicSettings = () => {
             description={<Translation id="TR_ONBOARDING_COINS_STEP_DESCRIPTION" />}
             innerActions={
                 <OnboardingCard.Button
-                    data-testid="@onboarding/coins/continue-button"
+                    data-testid="@onboarding/exit-app-button"
                     onClick={() => {
-                        goToNextStep();
+                        handleGoToSuite();
                     }}
                     isLoading={isTorLoading}
                     isDisabled={noNetworkEnabled}
